+++ Misc/NEWS.d/next/C API/2021-10-19-00-20-40.bpo-35134.Z0Zk_m.rst
@@ -1,0 +1,3 @@
+Exclude :c:func:`PyWeakref_GET_OBJECT` from the limited C API. It never
+worked since the :c:type:`PyWeakReference` structure is opaque in the
+limited C API.
